requirements.txt
# Core scientific computing
numpy=1.21.0
scipy=1.7.0
pandas=1.3.0

# Medical image processing
SimpleITK=2.1.0
pyradiomics=3.0.0

# Machine learning
scikit-learn=0.24.0

# Utilities
tqdm=4.62.0

# Optional for visualization
matplotlib=3.4.0
seaborn=0.11.0
nibabel=3.2.0